+++ src/svgMeasure.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var xmldoc = require('xmldoc');
+var isString = require('./helpers').isString;
 
 /** Strip unit postfix, parse number, but return undefined instead of NaN for bad input */
 function stripUnits(textVal) {
@@ -18,11 +19,11 @@ function parseSVG(svgString) {
 	try {
 		doc = new xmldoc.XmlDocument(svgString);
 	} catch (err) {
-		throw new Error('SVGMeasure: ' + err);
+		throw new Error('Invalid svg document (' + err + ')');
 	}
 
 	if (doc.name !== "svg") {
-		throw new Error('SVGMeasure: expected <svg> document');
+		throw new Error('Invalid svg document (expected <svg>)');
 	}
 
 	return doc;
@@ -31,17 +32,30 @@ function parseSVG(svgString) {
 function SVGMeasure() {
 }
 
-SVGMeasure.prototype.measureSVG = function (svgString) {
+SVGMeasure.prototype.measureSVG = function (svg) {
+	var width, height, viewBox;
 
-	var doc = parseSVG(svgString);
+	if (isString(svg)) {
+		var doc = parseSVG(svg);
 
-	var docWidth = stripUnits(doc.attr.width);
-	var docHeight = stripUnits(doc.attr.height);
+		width = doc.attr.width;
+		height = doc.attr.height;
+		viewBox = doc.attr.viewBox;
+	} else if (typeof SVGElement !== 'undefined' && svg instanceof SVGElement && typeof getComputedStyle === 'function') {
+		width = svg.getAttribute("width");
+		height = svg.getAttribute("height");
+		viewBox = svg.getAttribute("viewBox");
+	} else {
+		throw new Error('Invalid SVG document');
+	}
 
-	if ((docWidth == undefined || docHeight == undefined) && typeof doc.attr.viewBox == 'string') {
-		var viewBoxParts = doc.attr.viewBox.split(/[,\s]+/);
+	var docWidth = stripUnits(width);
+	var docHeight = stripUnits(height);
+
+	if ((docWidth == undefined || docHeight == undefined) && typeof viewBox == 'string') {
+		var viewBoxParts = viewBox.split(/[,\s]+/);
 		if (viewBoxParts.length !== 4) {
-			throw new Error("Unexpected svg viewbox format, should have 4 entries but found: '" + doc.attr.viewBox + "'");
+			throw new Error("Unexpected svg viewBox format, should have 4 entries but found: '" + viewBox + "'");
 		}
 		if (docWidth == undefined) {
 			docWidth = stripUnits(viewBoxParts[2]);
@@ -57,18 +71,28 @@ SVGMeasure.prototype.measureSVG = function (svgString) {
 	};
 };
 
-SVGMeasure.prototype.writeDimensions = function (svgString, dimensions) {
+SVGMeasure.prototype.writeDimensions = function (svg, dimensions) {
+	if (isString(svg)) {
+		var doc = parseSVG(svg);
 
-	var doc = parseSVG(svgString);
+		if (typeof doc.attr.viewBox !== 'string') {
+			doc.attr.viewBox = `0 0 ${stripUnits(doc.attr.width)} ${stripUnits(doc.attr.height)}`;
+		}
 
-	if (typeof doc.attr.viewBox !== 'string') {
-		doc.attr.viewBox = `0 0 ${stripUnits(doc.attr.width)} ${stripUnits(doc.attr.height)}`;
+		doc.attr.width = "" + dimensions.width;
+		doc.attr.height = "" + dimensions.height;
+
+		return doc.toString();
 	}
 
-	doc.attr.width = "" + dimensions.width;
-	doc.attr.height = "" + dimensions.height;
+	if (!svg.hasAttribute('viewBox')) {
+		svg.setAttribute('viewBox', `0 0 ${stripUnits(svg.getAttribute('width'))} ${stripUnits(svg.getAttribute('height'))}`);
+	}
 
-	return doc.toString();
+	svg.setAttribute('width', "" + dimensions.width);
+	svg.setAttribute('height', "" + dimensions.height);
+
+	return svg;
 };
 
 module.exports = SVGMeasure;